refactor: tests with mocked synapse server
