+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     restart: always
     user: postgres
     volumes:
-      - './postgres_data:/var/lib/postgresql/data'
+      - postgres_volume:/var/lib/postgresql/data
     environment:
       - POSTGRES_PASSWORD=postgres
       - POSTGRES_DB=social_service
@@ -18,3 +18,5 @@ services:
     user: redis
     ports:
       - '6379:6379'
+volumes:
+  postgres_volume:
