+++ src/entities/friendships.rs
@@ -105,13 +105,12 @@ impl FriendshipsRepository {
             Ok(rows) => Ok(rows
                 .iter()
                 .map(|row| -> Friendship {
-                    let friendship = Friendship {
+                    Friendship {
                         id: row.try_get("id").unwrap(),
                         address_1: row.try_get("address_1").unwrap(),
                         address_2: row.try_get("address_2").unwrap(),
                         is_active: row.try_get("is_active").unwrap(),
-                    };
-                    friendship
+                    }
                 })
                 .collect::<Vec<Friendship>>()),
             Err(err) => match err {
