+++ src/routes/v1/friendships/mod.rs
@@ -1,3 +1,3 @@
+pub mod errors;
 pub mod get;
 pub mod types;
-mod errors;
