--- tests/components/mod.rs
@@ -1,3 +1,0 @@
-pub mod database;
-pub mod redis;
-pub mod users_cache;
