--- tests/main.rs
@@ -1,5 +1,0 @@
-mod components;
-mod helpers;
-mod routes;
-mod v1;
-
