--- tests/routes/mod.rs
@@ -1,4 +1,0 @@
-pub mod check_auth;
-pub mod health;
-pub mod metrics;
-pub mod synapse;
