--- tests/v1/friendships/mod.rs
@@ -1,1 +1,0 @@
-pub mod get;
