--- tests/v1/mod.rs
@@ -1,1 +1,0 @@
-pub mod friendships;
